fix: server

+++ controllers/movies.js
@@ -1,7 +1,7 @@
 const Movie = require('../models/movie');
 const NotFoundError = require('../errors/not-found-err');
 const ValidationError = require('../errors/validation-error');
-// const OwnerError = require('../errors/owner-error');
+const OwnerError = require('../errors/owner-error');
 const CastError = require('../errors/bad-id-error');
 
 function getMovies(req, res, next) {
@@ -11,9 +11,23 @@ function getMovies(req, res, next) {
 }
 
 function createMovie(req, res, next) {
-  const { country, director, duration, year, description, image, trailer, nameRU, nameEN, thumb } = req.body;
+  const {
+    country, director, duration, year, description, image, trailer, nameRU, nameEN, thumb,
+  } = req.body;
 
-  Movie.create({ country, director, duration, year, description, image, trailer, nameRU, nameEN, thumb, owner: req.user._id })
+  Movie.create({
+    country,
+    director,
+    duration,
+    year,
+    description,
+    image,
+    trailer,
+    nameRU,
+    nameEN,
+    thumb,
+    owner: req.user._id,
+  })
     .then((movie) => res.status(200).send({ data: movie }))
     .catch((err) => {
       if (err.name === 'ValidationError') throw new ValidationError('Некорректные данные');
@@ -23,17 +37,19 @@ function createMovie(req, res, next) {
 }
 
 function deleteMovie(req, res, next) {
-  return Movie.findById(req.params.cardId)
+  return Movie.findById(req.params.moviedId)
     .then((movie) => {
       if (!movie) {
         throw new NotFoundError('Нет такого фильма');
       }
-    //   if (!movie.owner.equals(req.user._id)) {
-    //     throw new OwnerError('Вы не можете удалять карточки других пользователей');
-    //   }
+
+      if (!movie.owner.equals(req.user._id)) {
+        throw new OwnerError('Вы не можете удалять фильмы других пользователей');
+      }
+
       return Movie.findByIdAndRemove(movie)
-        .then((moviedel) => {
-          res.status(200).send({ data: moviedel });
+        .then((movieDel) => {
+          res.status(200).send({ data: movieDel });
         });
     })
     .catch((err) => {
